launchers: factor most of the code
